+++ release.nix
@@ -17,16 +17,43 @@ let
     ] ++ lib.optionals (reflex-platform.iosSupport) [
       "ghcIosAarch64"
     ];
-    hsPkgs = lib.genAttrs compilers (ghc: let
-      ghc' = reflex-platform.${ghc}.override {
-        overrides = self: super: import ./. self;
-      };
-    in {
-      inherit (ghc') reflex-dom reflex-dom-core;
+    variations = map (v: "reflex-dom" + v) [
+      ""
+    ];
+    compilerPkgs = lib.genAttrs compilers (ghc: let
+      variationPkgs = lib.genAttrs variations (variation: let
+        reflex-platform = reflex-platform-fun {
+          inherit system;
+          haskellOverlays = [
+            # Use this package's source for reflex
+            (self: super: {
+              _dep = super._dep // {
+                reflex-dom = builtins.filterSource (path: type: !(builtins.elem (baseNameOf path) [
+                  "release.nix"
+                  ".git"
+                  "dist"
+                ])) ./.;
+              };
+            })
+          ];
+        };
+        all = {
+          inherit (reflex-platform.${ghc})
+            reflex-dom-core
+            reflex-dom
+            ;
+        };
+      in all // {
+        cache = reflex-platform.pinBuildInputs "reflex-${system}-${ghc}-${variation}"
+          (builtins.attrValues all);
+      });
+    in variationPkgs // {
+      cache = reflex-platform.pinBuildInputs "reflex-${system}-${ghc}"
+        (map (a: a.cache) (builtins.attrValues variationPkgs));
     });
-  in hsPkgs // {
-    cache = reflex-platform.pinBuildInputs "reflex-dom-${system}"
-      (lib.concatLists (map builtins.attrValues (builtins.attrValues hsPkgs)));
+  in compilerPkgs // {
+    cache = reflex-platform.pinBuildInputs "reflex-${system}"
+      (map (a: a.cache) (builtins.attrValues compilerPkgs));
   });
 
   metaCache = native-reflex-platform.pinBuildInputs "reflex-dom-everywhere"