+++ reflex-dom/ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for reflex-dom
 
+## Unreleased
+* On Android, enable propting the user for geolocation
+  permissions on demand by default. See
+  https://developer.android.com/reference/android/webkit/WebChromeClient.html#onGeolocationPermissionsShowPrompt(java.lang.String,%20android.webkit.GeolocationPermissions.Callback)
+  for details.
+
 ## 0.5.1
 
 * The default jsaddle backend on macOS when built from nix